Refactor code to align types between layers
Change types more appropiated to the value will hold those variables
Setup test and app database properties
Minimum test to check the retrieval of data using JPA repositories

+++ src/main/java/org/javierperis/pricing/PriceDsRequestModel.java
@@ -7,7 +7,7 @@ import java.time.LocalDateTime;
 @Value
 public class PriceDsRequestModel {
 
-    Long brandId;
+    Integer brandId;
     Long productId;
     LocalDateTime date;
